requirements.txt
fastapi
uvicorn[standard]
pillow
google-genai
python-dotenv
